move athene endpoint to public
