+++ src/main/java/de/tum/in/www1/artemis/web/rest/publicc/AtheneResource.java
@@ -1,4 +1,4 @@
-package de.tum.in.www1.artemis.web.rest;
+package de.tum.in.www1.artemis.web.rest.publicc;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,7 +16,7 @@ import de.tum.in.www1.artemis.service.connectors.athene.AtheneService;
  * REST controller for managing Athene results.
  */
 @RestController
-@RequestMapping("api/athene-result/")
+@RequestMapping("api/public")
 @Profile("athene")
 public class AtheneResource {
 
@@ -32,15 +32,14 @@ public class AtheneResource {
     }
 
     /**
-     * Saves automatic textAssessments of Athene
+     * POST athene-result/:exerciseId -- Saves automatic textAssessments of Athene
      *
      * @param exerciseId The exerciseId of the exercise which will be saved
      * @param atheneResponse The calculation results containing blocks and clusters
      * @param auth The secret for authorization
      * @return 200 Ok if successful or 401 unauthorized if secret is wrong
      */
-    // TODO: /public
-    @PostMapping(value = "{exerciseId}", consumes = "application/x-protobuf")
+    @PostMapping(value = "athene-result/{exerciseId}", consumes = "application/x-protobuf")
     @EnforceNothing
     public ResponseEntity<Void> saveAtheneResult(@PathVariable Long exerciseId, @RequestBody AtheneResponse atheneResponse, @RequestHeader("Authorization") String auth) {
         log.debug("REST call to inform about new Athene results for exercise: {}", exerciseId);
